third_party/WebKit/LayoutTests/fast/css/recalc-inherit-001-expected.txt

Check that we don't do unnecessary style recalcs for inheritance.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS successfullyParsed is true

TEST COMPLETE
